+++ backend/src/main/java/com/api/backend/repositories/TransactionRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.api.backend.models.Transaction;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
     @SuppressWarnings("null")
-    Optional<Transaction> findById(Long id);
+    Optional<Transaction> findById(Integer id);
 
     Transaction deleteByUser_id(UUID userID);
 }
